refactor: arrange `RedirectLink` into `components` folder
