+++ lcid-frontend/src/components/RedirectLink.jsx
